+++ Tests/DangerTests/BitBucketServerTests.swift
@@ -162,4 +162,22 @@ final class BitBucketServerTests: XCTestCase {
         XCTAssertEqual(user.type, "NORMAL")
         XCTAssertNil(user.emailAddress)
     }
+
+    func testItParsesTheBitBucketPullRequestFromForkedRepo() {
+        let bitBucketServer = bitbucketForkedRepoFixtureDSL.bitbucketServer
+        let pullRequest = bitBucketServer?.pullRequest
+
+        let expectedProject = BitBucketServer.Project(id: 1, key: "PROJ", name: "Project", isPublic: nil, type: "PERSONAL")
+        let expectedRepo = BitBucketServer.Repo(name: "Repo",
+                                                slug: "repo",
+                                                scmId: "git",
+                                                isPublic: false,
+                                                forkable: true,
+                                                project: expectedProject)
+        let expectedBase = BitBucketServer.MergeRef(id: "refs/heads/master",
+                                                    displayId: "master",
+                                                    latestCommit: "8942a1f75e4c95df836f19ef681d20a87da2ee20",
+                                                    repository: expectedRepo)
+        XCTAssertEqual(pullRequest?.fromRef, expectedBase)
+    }
 }